Phase K Wave 59: 0x00BC CTA_REQUEST → 0x00BD CTA_RESPONSE direct-reply round-trip + sizeof(long) WRITE-path tightening
Adds a direct-reply round-trip integration test for the CTA_REQUEST →
CTA_RESPONSE pair: send 12B {SourceID=0x12345678, TargetID=0, Action=0}
at sectorId=10151 (Luna Station handshake), assert a 9B reply that echoes
SourceID + Action and ends with Success=0x01.

Action=0 forces PlayerManager::GroupAction's default arm — a logging
no-op that never dereferences SourceID/TargetID, so the request is safe
with arbitrary values and no group/party seed.

Server-fidelity tightening at PlayerConnection.cpp:7740-7741:
two *((long*) &CTAResponse[N]) writes into a 9-byte stack array on
Linux x86_64 (sizeof(long)==8) wrote 16 total bytes instead of the
retail-correct 8, with the second write's high 3 bytes spilling past
the end and clobbering byte[8] (Success) with the sign-extended high
byte of Action. Switched to *((int32_t*) ...) — pure fidelity
restoration: on Win32 sizeof(long)==4, the originals were
byte-equivalent. Authorized by CLAUDE.md: "It does NOT restrict changes
that tighten the server toward greater fidelity ... those are always
welcome."

Surfaces the third leg of the Phase K sizeof(long) bug taxonomy:
  Leg 1 (Waves 7/11/12): READ over-read past wire-payload end
  Leg 2 (Waves 11/12): struct-layout shift on long widening
  Leg 3 (Wave 59): WRITE overflow clobbering adjacent stack field

+2 coverage ratchet (92/207 → 94/207, 45.4%) — third +2 wave of Phase K.

Touches:
- server/src/PlayerConnection.cpp:7740-7741 (long→int32_t + WHY comment)
- tools/cli-client/src/CliClient.Core/Opcodes/OpcodeId.cs (+CtaRequest 0x00BC, +CtaResponse 0x00BD)
- tests/integration/CliClient.IntegrationTests/Opcodes/SectorCtaRequestTests.cs (NEW)
- tests/integration/CliClient.IntegrationTests/Coverage/TestedOpcodes.cs (MinTestedCount 92→94, +2 entries)
- tests/integration/CliClient.IntegrationTests/Fixtures/seed.sql (+cli_test56)
- tests/integration/CliClient.IntegrationTests/TestAccounts.cs (+test→cli_test56)

Test passes (7s); coverage ratchet test passes (94/207).

+++ server/src/PlayerConnection.cpp
@@ -7737,8 +7737,13 @@ void Player::HandleCTARequest(unsigned char *data)
 		0x01						//Success
 	};
 
-	*((long*) &CTAResponse[0]) = myCTARequest->SourceID;
-	*((long*) &CTAResponse[4]) = myCTARequest->Action;
+	// Phase K: wire fields are 4-byte each (Win32 sizeof(long)==4 so the
+	// original long* writes were correct on retail; on Linux x86_64
+	// sizeof(long)==8 made each write 8 bytes, overflowing the 9-byte
+	// CTAResponse buffer and clobbering the Success byte (offset 8)
+	// with the high byte of the sign-extended Action int32.
+	*((int32_t*) &CTAResponse[0]) = myCTARequest->SourceID;
+	*((int32_t*) &CTAResponse[4]) = myCTARequest->Action;
 
 	SendOpcode(ENB_OPCODE_00BD_CTA_RESPONSE, (unsigned char *) &CTAResponse, sizeof(CTAResponse));
 }
